query: Fix bug where empty quoted key wasn't accepted
These are perfectly valid in TOML:
  https://toml.io/en/v1.0.0#keys
so we always intended to accept them in queries.

We hadn't been, due to a bug in `nom`: its `escaped_transform`
fails when the matching string is empty, unless it happens to be
at the end of the input.

(Certainly sometimes one wants to reject an empty string, and one
could imagine wanting `escaped_transform` to do so.  But its docs:
  https://docs.rs/nom/7.1.1/nom/bytes/complete/fn.escaped_transform.html
don't indicate any special behavior for the case of an empty string,
so doing so is a bug.)

The bug was already present on the old `nom` version we used to use,
and remains present at 7.1.1, the latest.

Fix the issue for us by working around the bug in `nom`.

Fixes: #20