tools: Make build-release more reproducible
In particular, this should mean that a given Git commit always
produces an identical tarball, provided it produces an identical
`toml` binary.

And the latter empirically holds when I just rerun the build on my
own machine.  I expect it will hold as long as one uses the same
version of the Rust toolchain.